requirements.txt
astroid==2.2.5
backcall==0.1.0
bcrypt==3.1.6
blinker==1.4
cffi==1.12.3
Click==7.0
colorama==0.4.1
decorator==4.4.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.17
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.3.1
pylint==2.3.1
six==1.12.0
SQLAlchemy==1.3.3
traitlets==4.3.2
typed-ast==1.3.4
wcwidth==0.1.7
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
